Moved check_reference_file() to utils/.
